+++ src/main/java/pl/telvarost/mojangfixstationapi/client/skinfix/SkinService.java
@@ -104,7 +104,7 @@ public class SkinService {
         try {
             if (profiles.containsKey(name)) return;
 
-            if (Config.ConfigFields.prioMojangProvider){
+            if (Config.ConfigFields.prioritizeMojangProvider){
                 providers = Arrays.asList(new MojangProfileProvider(), new AshconProfileProvider());
             } else {
                 providers = Arrays.asList(new AshconProfileProvider(), new MojangProfileProvider());